Remove incorrect assert to stop from molding HALT/etc.
This assertion is a garbled holdover of an attempt to try and not
mold out a "special" ERROR! type, such as the internal ones
used to implement halt and PARSE's longjmp hack.  But the sign
on the comparison is wrong, and moreover it should probably
be able to work anyway.  (Mold is used often in debugging cases,
so there are much better places to check to ensure that such
errors haven't leaked into userspace.)