Create "padding" categories for unused errors
Behind the scenes, Find_Error_Info() was using some math to
know specifically about the number of categories in errors.r.
When categories went missing, the user error (numerically at
1000) wasn't being found, because it was in the slot expected
for category 7.

Adds padding categories and a comment regarding why.

Also links the submodule to the version of the test repository
with updates related to errors.