Add AND? OR? XOR? NOT? operations
This is a step toward CC#1879 moving AND, OR, and XOR to be
conditional, by providing the conditional forms of the operations
in prefix form. The baseline version of NOT was renamed to NOT?
o fit in with the group, and then aliased to NOT (since it has no
infix form).