Differentiate EVERY return result, loop constants
The name EVERY is used in underscore.js for a routine that
will test a condition for every item in a series.  Although the
idea was to use EVERY as a synonym for FOR-EACH, it
seems that having a cumulative return result that's more like
ALL MAP-EACH is more useful and fitting for the name EVERY.

This adds a starter implementation for EVERY, but note that it
is subject to a quirky behavior regarding UNSET! which is to
treat it as if it were conditionally true.  That is the case for ALL
right now as well.  Both should be resolved, but that is a
separate issue.

There were also some odd numeric constants used in the
loop code, and this goes ahead and turns those into an
enumeration.