Simplify function create/copy, document, fix hazard
The static analyzer was complaining about something in the
function creation code, which called attention to it having a few
half-baked features that were dangerous and weird.  MAKE could
take a function value (instead of e.g. MAKE FUNCTION!, so you
could write MAKE :APPEND for instance) and then you could either
overwrite the spec or the body or both.

This is a feature which is being considered for implementation in
a safe way, but the unsafe way was allowing the creation of specs
that were incompatible with the body.  It used an ad-hoc syntax of
putting an asterisk in the spec slot if you wanted to reuse it, when
it would be possible to avoid syntax by merely putting references to
the functions you wanted to borrow from in the spec or body slots.

The only use of the feature was to make EMPTY? re-use the code
body of TAIL?, so it is GC'd along with comments explaining what
is there more simply.  The feature will be added, just not like this.