Error on uninitialized year code path of Scan_Date
It wasn't immediately obvious to me how to trigger this branch,
but if it is triggered then it would lead to using an uninitialized
value for year.  As the function is able to return FALSE on
invalid input, that seemed a safer response.  Should anyone
trigger it as an error, they can drive the issue of how to
generate a correct year for that case.

Found by clang static analyzer.