Eliminate PF_CASE/PF_CASED ignored flags
Clang static analyzer introduced me to a recent bug introduced
regarding the binary case sensitivity flag not being picked up
correctly...which also led me to notice another bug that was
longstanding relating to an inconsistent use of the "flags" field
of the parse state.  It was being OR'd with the flags for a FIND
search, yet it tried to inject its own flag numbering scheme for
PF_ flags that weren't actually part of the main parsing flags
enumerated type.

To make what was going on clearer, I renamed the flags passed
to find `find_flags` and eliminated the PF_CASE and PF_CASED
flags from the superflous enum.

Props again go to clang static analyzer.