Clarify *-EACH loop code, static analysis warning
The static analyzer does path analysis and it can't necessarily
tell by looking at two paths controlled by different conditions
whether those conditions will return the same result and thus
mean a variable initialized in the code under the first will be
good in the second.  You can hint it by assigning the thing you
are testing to a constant, so then it knows it won't change.

This also updates the common loop code for the MAP-EACH,
FOR-EACH, etc. functions so that it refers to the variable you
are iterating over as "data" instead of "value" (which is the
name of the parameter in the function spec, also).