Add error on invalid argument code path of trim
There are several Rebol natives that take for granted that the
types were already checked, so e.g. if they know they are
getting an integer or string they can say `if (IS_INTEGER(...))`
and `else if (IS_STRING(...))` and then have no additional else
to handle the case where neither is true.

The options are either to add an else branch that errors, or to
remove the `if` from the last `if else`.  This removes the if and
adds an assert, which helps communicate the intention better
and micro-optimization-wise winds up slightly faster in a release
build.