Hint static analyzer about varargs in routine
As mentioned in a prior commit, the static analyzer does not
know that conditions controlling two code paths will necessarily
be returning identical results.  So if an assignment runs in a
first conditional that is depended on by a second, it may see
the second as use of an uninitialized variable.

This introduces a constant (optimized out) which helps clue
the analyzer that the controlling condition of two different IF
statements is actually the same condition.