Change to function that makes scan errors without raising
The scanner has a "relaxed" mode in which it does not raise
errors, but returns them to a caller.  This was implemented by
a function that took an error target, tested it for NULL, raised
it if it was or write the error into it if it wasn't.

Besides being a bit convoluted, the name "Scan_Error" for this
function fit into the forms of other "Scan_XXX" functions whose
job it is to scan a value of XXX type, which makes it seem like
Scan_Error might be used to take an input string and give you
back an ERROR! value.

This clears both issues up with a function that just makes the
error value that has "Make" in the name, and then lets the caller
take care of whether to raise it or not.