Eliminate Scan_Code routine (use Scan_Block instead)
The routine Scan_Source was described as being used to "Scan
source code", and it was a convenience wrapper that took care of
the Init_Scan_State for you.  It then called the routine Scan_Code
(also used by TRANSCODE) which was the same but you could
specify more options by doing one's own Init_Scan_State call.

The Scan_Code routine's name is difficult to distinguish from
Scan_Source (and also looks as if it returned some kind of scanning
code number).  It was identical to Scan_Block, except Scan_Block
can be called recursively...so Scan_Code had to clean up any
lingering emitted values remaining from a previous scan that had
trapped before being able to process its emits.

This moves the cleanup into Init_Scan_State and replaces the calls
to Scan_Code with Scan_Block.  Also, since there were no clients
of the "zero means call the string length function" protocol for
Scan_Source, that was eliminated (as that idea is getting the axe in
general around the codebase).