Adjust TOKEN_XXX/LEX_XXX naming
In trying to figure out what the TOKEN_XXX were, I observed that
TOKEN_BLOCK pairs with TOKEN_BLOCK_END, and as such it
does is better seen as TOKEN_BLOCK_BEGIN.  This does a
similar adjustment for TOKEN_PAREN and makes TOKEN_LINE
more clearly a newline token.

There was also a need in the code to make explicit what kind of
"quote" a LEX_DELIMIT_QUOTE meant, so I distinguished that to
be LEX_DELIMIT_DOUBLE_QUOTE (it doesn't occur often enough to
need an abbreviation) and changed LEX_SPECIAL_TICK to just be
LEX_SPECIAL_APOSTROPHE, because TICK is not a standard term and
could also pertain to ticks for TIME! or something else.