Comment Locate_Token, clarify BUF_MOLD
The Scan_XXX family of functions generally take in a prepped
Scan_State and load its contents into a value.  However,
Scan_Token was "special" and really just went and found the start
and end pointers...with the value loaded by a later routine.  Locate
seemed the better term to help distinguish its function.

This adds a lengthy comment and some examples of what the function
seems to do.  It also makes the hidden dependency on BUF_MOLD more
obvious, because that was called out in the comments.  This will
likely need to be evolved into a better design.