Clarify what Prescan and classifier macros do
The Prescan() function was very weird, and needed a better
explanation.  Changed the name to Prescan_Token() to convey
more that it is a special purpose partner to Locate_Token().

There were several macros that were used with the LEX_MAP
which spoke in terms of the relative values of the lexical category
(e.g. IS_LEX_AT_LEAST_SPECIAL(...)).  This implied some sort
of ranking...which there isn't.  It was a trick for using the relative
positions in the enum to do a quicker comparison without having
to OR together multiple bitmasks.  This changes the names to
the thing they actually test for (e.g. IS_LEX_NOT_DELIMIT is
what IS_LEX_AT_LEAST_SPECIAL actually tested).