NOT_NEWLINE => !ANY_CR_LF_END, END delimit name
Reading code for the NOT_NEWLINE macro shows that it tests for
whether something is a CR, LF, or a null terminator to signal the
end of string.  NEITHER_CR_NOR_LF_NOR_TERMINATOR was
a pretty long form of negative logic to replace it with, so turning it
to positive logic and then negating it at the callsite made more
sense.  Although !ANY_CR_LF_END(...) is longer it is more
descriptive of what is actually happening, which is very relevant
to the code when one is reading it...because the routine has to
process a CR, LF, or END after the instruction runs.

This also changes the "end of file" designation of 00 to just END,
because it's really an end of string byte...you can scan source
from a byte string with no file at all.  Calling it TERMINATOR was
also possible but it would've been longer.

There were some missing parentheses on IF and WHILE which
were working because the macros they called had parentheses
in them.  If there are any more of those, they should be hunted
out in another commit.