Make TAIL? spec support all types EMPTY? does
I failed to notice that EMPTY? had a less-constrained type spec
when I was getting rid of the single instance of native-guts-reuse
in Make_Function, even though A_TAILQ is used to implement
TAIL? which was used to implement EMPTY?.  :-/  Whether this
is important or not is subject to debate, but the TAIL? spec
definition says:

{Returns TRUE if series is at or past its end; or empty for other types.}

So it seems to be setting itself to be a synonym for EMPTY?
already with the "for other types" designation.  Hence it seems
adding OBJECT! and NONE! as acceptable input to TAIL? is an
all right way to work around this for now.  This may just wind up
being a stopgap measure, but the alternative of making the action
A_EMPTY and then implementing TAIL? in terms of EMPTY?
would be more churn that doesn't seem very well-motivated.