Move eval from DO to EVAL, <r3-legacy> mode
This adds the special EVAL primitive, which can handle a value inline
however the evaluator would:

    >> eval (quote x:) 10
    >> print x
    10

    >> eval :add 10 20
    == 30

What makes it special is that it is not a function that could be written
in user code, because it has an arity dictated by its argument...and
user functions have to be fixed arity.

Previously DO was the only operation that exposed the evaluator in
this special way, but it only worked with functions.  EVAL pulls that
out of DO (making it a normal function) and gives the behavior for
al data types, which localizes the specialness.

With DO more limited, it creates another divergence from R3-Alpha,
so this commit carries a tool in it with a new behavior for how DO
handles tags.  You can say DO <r3-legacy> and get legacy behaviors
back...such as making DO accept more types again, or putting back
simple parse, etc.  Older Rebols will treat this as a no-op because
DO of a tag didn't do anything, which makes it a good method.

The one thing that can't be put back--however--is for the wrapped
DO function to behave like EVAL.  This is because the operation is
special...and since you cannot write a function that acts like EVAL
does, you cannot write a wrapper for it either.  So basically, any DO
calls on functions *must* be changed to EVAL.  The R3_LEGACY
environment variable controls paying attention to do-runs-functions
as a small compromise.

DO of a tag long-term is conceived as a way to invoke a named
module picked out of some directory service.  This commit adds a
few hardcoded items as a starter set (though not all of them are in
working order to be loaded from URL by Ren/C yet, it's a call to
action to make it so...)