Friendlier handling of NONE/STRING parse rules
The C code for "simple parse" had been deleted, and the function
interface changed to not take NONE! or STRING!.  Yet it turned
out that simple parse was a frequently used feature.  In order to
make it easier to know what to do (change to use SPLIT) this
adds in an error message saying so.

Additionally, if you use DO <R3-LEGACY> then this puts in a
version of PARSE that actually will support simple parse by
making the call to SPLIT itself.