CHANGE-DIR to URL, relative paths for DO of URL
This adds support for the ability to change-dir to a URL.  It does not
add support for listing the contents of URL!s...and there is no
common convention for HTML directory listings.  However there may
be something to put in port schemes for SFTP and others (even
HTTP could get plugins for scraping out data for GitHub directory lists,
or auto-generated pages by apache for servers with that enabled).

The motive for this was in particular to facilitate a DO of a URL, in
which that source has relative paths via file to `do %something.reb`.
By letting the DO of the URL put Rebol "into that path", then when
a FILE! is run it will be relative to that URL.  This way, if you have
files organized on web server paths to match the scheme you
would have them locally, it will still work.