Conditional AND/OR/XOR (#1879)
This makes the undecorated AND, OR, and XOR operations
"conditional" as opposed to bitwise.  They are the infix forms
of the prefix operations AND?, OR?, and XOR?.

Though #1879 describes a belief that the prefix forms of the
AND, OR, and XOR operations could be merged into
INTERSECT, UNION, and DIFFERENCE...that does not seem
to actually give full coverage.  For instance, when a BINARY! is
considered as a set of bytes, the effect of a union is different
from if it is considered as an array of bits to be OR'd together.

Hence this commit leaves the operations separate.  The infix
form of AND is changed to AND* to reflect its "boolean multiply"
algebraic equivalent (and suggest a mathy output as opposed
to a logic one).  Similarly for OR+ and XOR-.

It's possible to configure Rebol back to the bitwise defaults with
`do <r3-legacy>`.  Also, r2r3-future.reb is updated with some
help to create the operations...but since older Rebols cannot
make new infix operators they cannot be made fully compatible.