Properly handle date scan case from static analysis
Clang static analyzer noticed something suspicious in the date
scanning for a date type I didn't know examples of with a leading
year.  I didn't know if this format was in the wild or not, so I made
it not be handled before doing closer inspection.

@johnk found an instance of the date format:

    2009/04/20/19:00:00+0:00

Which probably isn't a good format to support in source due to
the slashes (and could be delegated to MAKE DATE! at the least)
But in any case, it turns out upon looking closer that the code path
where the uninitialized year is made is guarded by a test of the
day...this makes that more clear and suppresses the static analysis
warning using a dummy assignment.