Sweeping RETURN-capture fixes using <transparent>
This fixes a bunch of mezzanines which unwantingly captured RETURN (and
EXIT). This was because of the missing functionality to create functions
transparent to RETURN, which was achieved using the [throw] attribute in
R2 and is now again possible as a first step towards definitional RETURN
using the <transparent> attribute.

This commit fixes the following mezzanines/intrinsics:

- COLLECT
- DP, DELTA-PROFILE
- DT, DELTA-TIME
- FIND-ALL
- IN-DIR
- MAKE module!
- MODULE
- OBJECT
- USE
- WRAP

This, together with the implementation of <transparent> in commit
fe7762, fixes CureCode issue 539 (at least significant parts of it).