Legacy switches available in all debug builds
The evolving story for how to continue running code in Ren/C
which has not been updated (either partially-or-fully) is that DO of
<r3-legacy> will turn back the clock and morph the behavior to act
more like an R3-Alpha.  This requires some shim code written as
Rebol functions working along with some conditional behavior in
the C code.

The idea originally was that these switches should be hidden and
very hard to get...controlled through an environment variable so
you wouldn't accidentally be running them.  But as the idea
emerged that they'd only be in debug builds, that seems like
enough of a disincentive to depend upon it without making it any
more difficult than that.