Get symbol ID from type enum more intentionally
The %words.r file is set up in an order where the words representing
data types are at the beginning of the file.  There was an enigmatic
addition of 1 to the enum value (because 0 is reserved for "no symbol"
and it is REB_END the type, so END! is at position 1).

Since this addition was done in several places, this introduces a
macro for getting the symbol from a datatype...so that if the way to
get the symbol ever changes it can be updated in one place.