Test GROUP! compatibility alias for PAREN!
Renaming types in Rebol has technical barriers that are more like
renaming a dialect keyword than a function.

This gets the process started on a GROUP! rename by first testing
the compatibility tricks which would be used for PAREN! on it.
That means GROUP! is the compatible type, while PAREN! is the
"real" datatype. The fallout can be studied for what can be made to
work or not, and make it better before the switch is flipped to bias
it the other way where PAREN! is the alias and GROUP! is the
actual type.

This is spurred by realizing that the TYPE?/WORD call is a good
place to leave in compatibility code for returning PAREN!. But now
with the feature of being able to compare type names and words
loosely that people will be tempted to use the feature...dropping
the /WORD refinement. If that happens before any type renames,
then it makes it harder to keep compatibility.

So logistics of any type renames needs to get a started sooner
rather than later. This will give some experience with it.