Move /WORD off of TYPE-OF and onto TYPE?
This gets rid of /WORD on the main TYPE-OF native, which was not
really necessary because you could always write TO-WORD TYPE? but
it was considered cleaner to say TYPE?/WORD.  The TYPE? function is
given a /WORD refinement for compatibility purposes.

One good reason for losing the refinement is the ability to lax-compare
a word against its type from 9745502 (strict comparisons will still not
consider any datatype! equal to any word!).  Unfortunately, until that
setting cannot be turned off, any mezzanine or system code which has
to run with legacy code must assume the feature is not available.  Hence
TO-WORD (or TO WORD!) is still necessary several places, but marked
with OPTIONS_DATATYPE_WORD_STRICT to be cleaned up when that
is deemed no longer necessary.

Something emergent from this is that the datatype equality test and the
TYPE?/WORD legacy point present an opportunity window for changing
the names of types which needs to be focused on sooner rather than
later.  As long as TYPE?/WORD references exist in code that uses the
old name of a type, it can give back the old type while new written clients
go along happily with the new names.  e835358 summarizes the issue.