Have Make_Binary and Make_Unicode terminate
In order to optimize the number of writes, Rebol would not write the
terminator for series at allocation time...because it would be wasted
work if the series was going to just overwrite the terminator with
data and have to terminate again.

The responsibilities in the protocol are not completely clear, and
are awaiting formalization.  As an example: the Make_Binary and
Make_Unicode routines take a length and then terminate, which
assumes that the caller will then fill the data in from the start up
to the length.  However, some clients considered the length to
be a capacity and expected to be getting a terminated empty
series that they did not fill.

To patch the issue, this errs on the side of caution and terminates
both intents.  A more systematic answer about how termination
is handled in series is needed...so this just a minimalist band-aid
until that answer is codified.