Fix strerror_r usage on OS/X
.. and potentially on other POSIX-compliant platforms as well.

The code resolving a C-level error into an error message uses the
thread-safe strerror_r variant. Unfortunately, historic versions of
GNU's libc (glibc) provide a strerror_r variant with a different
signature, predating the standardisation as part of POSIX.1-2001.

The pre-processor feature test we use to discern between the two
strerror_r variants was taken directly from glibc's documentation.
Taking a glibc point-of-view this feature test focused at detecting when
the non-GNU variant was provided. Building on that, we used the GNU
variant, unless we explicitly detected the POSIX variant. However, other
platforms provide the POSIX variant as well, so this commit switches the
feature detection around to always use the POSIX-compliant strerror_r
unless we know the GNU variant is provided.

The cause for this change was OS/X's libSystem, which provides a
POSIX-compliant strerror_r as well, thus causing the following
compilation error (with -Werror):

    ../src/os/posix/host-error.c:146:8: error: incompatible integer to
          pointer conversion initializing 'char *' with an expression of
          type 'int'
          [-Werror,-Wint-conversion]
            char *maybe_str = strerror_r(errnum, str, len);
                  ^           ~~~~~~~~~~~~~~~~~~~~~~~~~~~~

To deteriorate matters, _GNU_SOURCE was _always_ defined manually
(except for C++ builds) as a left-over of the host sources
restructuring: the elaborate feature testing was accidentally short-cut.
This commit remedies that as well.

One important thing to note, from glibc's documentation:

> "If no feature test macros are explicitly defined, then (since glibc
> 2.4) _POSIX_SOURCE is defined by default with the value 200112L, so
> that the XSI-compliant version of strerror_r() is provided by
> default."

So most likely, the previous feature test was only implemented as a
side-effect of accidentally defining _GNU_SOURCE for this file (hence
forcing the fallback to the GNU strerror_r implementation).

Even if accidentally defining _GNU_SOURCE is not the root cause, we are
now doing this test dance for supporting pre-glibc-2.4 builds. Glibc 2.4
was released in 2006. Distros coming with pre-2.4 glibc are for example:

- Debian 4.0 "etch": initially released 2007, last point release 2010,
  discontinued in 2010
- Ubuntu 6.06 LTS "Dapper Drake: initially released 2006, last point
  relase 2008, end-of-life 2011
- Red Hat Enterprise Linux 4: initially released 2005, last point
  release 2011, extended life-cycle support until 2017.