ANY- prefix for SERIES!, NUMBER!, SCALAR!
When ARRAY was picked to be the general category word for BLOCK,
PATH, and GROUP...it opened up the ability to make the typeset and
type check either have an ANY- on them or not.  (It did not have that
opportunity when BLOCK was already taken for a category member.)

That opportunity landed on the side of thinking the ANY- was a good
cue for those reading the code.  It helps educate new users to realize
they are not dealing with a concrete type.  It also sort of "self-checks"
because one would not think to write MAKE ANY-ARRAY! because
that comes off as obviously nonsensical, whereas one might have a
harder time immediately processing why MAKE ARRAY! gave an
error after several ARRAY? tests seemed to work just fine, in the
context of a routine with several usable ARRAY! parameters.

This drew scrutiny on the three type sets without ANY- prefixes on
them: NUMBER!, SERIES!, and SCALAR!.  After seeming universal
consensus that ARRAY! should have the ANY-, putting it on those
typesets makes sense for the same reasons.  It's also not that
disruptive in terms of compatibility as typeset aliases work nearly as
well as typesets, since there's no TO-WORD conversion for them.

(The actual word used by the system has one advantage, that MAKE
TYPESET! will accept that word unbound...but if bound, the word
will be looked up to the type or typeset it references.  I know this
is there because I looked, but I don't know if anyone uses it.  Adding
that particular compatibility switch can be done on-demand if so.)

Commit also (slightly) tidies how typesets are made during boot.