Use custom HAVE_EXECINFO_AVAILABLE to enable crash backtraces
There seems to be no portable way to detect the presence of the
"backtrace" GNU extension. Therefore we resort to only enabling it via a
custom HAVE_EXECINFO_AVAILABLE define. As we currently do not have a
generic build-configuration facility, defining HAVE_EXECINFO_AVAILABLE
currently must be done manually.

Also, while we are at it, headers included in src/os/posix/host-error.c
are cleaned up.