Do not skip object/func frames during GC
There was a performance-based exemption for "word lists" (actually
"unwords") in function and object frames.  Because these words did
not have actual bindings and were just bitsets, they were taken out
of the consideration for GC.

This isn't a very forward-looking way to optimize.  Firstly, because a
richer typeset model may not be limited to just 64 bit flags for the
built in types, but might have series of object values representing
user-defined types.  Also, if symbols themselves are ever going to
be GC'd (they should be) then the symbol ID for the parameter
will need to be considered.

A more global optimization would be if there was a flag that was
maintained about whether there might be any GC'able values in
an array.  It could start out saying there may be...but then if it did
a visit and didn't see any mark it as not needing GC.  Then
modifications dirty that bit.