Get rid of UTYPE! stub, datatype=>symbol fix
This gets rid of UTYPE!, though it's not really getting rid of much
because UTYPE! had never been implemented.  Current ideas about
how to get user-extensible datatypes are planning to build upon the
object system and improve that instead.

Trying to delete the type led to a crash where the conversion
between datatypes and symbols had an off-by-one error based on
the + 1 / - 1 tapdance done to avoid taking the 0 symbol.  This had
already gotten a start on having macros to point out where this was
being done, so this just adds a couple more of those macros to
cover the crashing case.  It also updates one spot to use them (any
others that are found should be modified as well).

Because removing the t-utype.c file from the project requires doing
`make make`, it can be frustrating to go back and forth to past
versions when that has to run each time.  So this doesn't delete
the file yet, but will do so when a batch of necessary changes to
the makefiles comes up.