Make SAME_SYM take SYMs, not VALUEs
There are supposedly different symbol and canon processes
for words and "unwords", where VAL_WORD_CANON/SYM is
different from VAL_BIND_CANON/SYM.  There doesn't actually
seem to be a difference in implementation of the macros, but if
there was then the it's being gotten wrong in MAP!.

As it so happens... with the upcoming elimination of UNWORDs,
the two actually *will* be different, and so it will matter.  This fixes
the issue with map and simplifies SAME_SYM so it can be used
on symbol #s directly instead of values.