SYM_NOT_USED => SYM_0, strengthen asserts
Because the SYM_ pattern for constants covers names of Rebol
words, SYM_NOT_USED seems like it could be the enum value
corresponding to `not-used` the WORD!.  SYM_0 is a good
replacement because it is not likely the system would need a
WORD! non-natural for zero.

This commit also enforces the invariant that Val_Init_Word won't
take a SYM_0 value, which involved some slight change to the
scanner so that it creates the word all in one go instead of
making it with a 0 symbol and then poking the symbol in after.