Add assert on lookup that frame symbol matches
An object or function frame contains a symbol, and lives at a certain
index.  This index gets put into bound words along with the frame
pointer when a binding happens.

There wasn't any assertion in the system that when you went to look
up the frame key at an index that the symbol matches.  It seemed a
prudent one to add, when shuffling the keys from being an "UNWORD"
to being typesets...though it may be worthwhile to keep around for
a while even after that settles.