Change Get_Action_Word to Get_Action_Sym
Get_Action_Word used to return an UNWORD out of the frame
keys directly.  When leaked into the system they could cause crashes
if the encoded type information were misinterpreted as binding info.

Also, in the pending change where TYPESET! takes the place of an
UNWORD, if it didn't create a new REBVAL you'd get a typeset...