Main unword-to-typeset conversion commit
This commit does the main work after the "pre-renaming"...for ease of
inspection.  It still has another "post-renaming" for those renames
that only make sense after the change.  Summarizing:

Historically there has been a special variation of WORD! used to store
the keys of object frames and the parameters+locals of a function.  It
had in common that in the structure it had the same symbol number
field that regular words did, but instead of binding information it held
a typeset.  It held this 64-bit typeset struct field at the same offset that
actual TYPESET!s did, so functions could run in common.

Experience bore out that this was both confusing and dangerous.  The
unword would leak easily to code that wasn't sensitive to its special
status, and crash using type bits as binding.  People reading code and
trying to follow it would not be cued very well to knowing that what they
were looking at was not "a word as they knew it".

The pressing reason for the change was contemplating enhancements
to typesets so they could expand out to more than just 64 flag bits of
information on types (which is a nice optimization for common cases
but not forward-looking in a future with user-defined types).  It makes
sense to think of the typeset vocabulary and mechanics growing...and
in that point of view it makes more sense to see these keys as "typesets
with a word symbol ID in them" as opposed to "words that can't bind and
have typesets in them".

(Also: some things, like a <protected> status, makes sense being in the
representation of the typeset more than being a flag inside of a word.
For instance there's how `const` is part of the datatype in C.)

One issue from having it be switched to a "typeset plus symbol" is that
the frame keys lose the SET, GET, LIT, REFINEMENT distinction.  So
the properties those distinctions made are part of the typeset EXT bits.