Post-unword elimination renamings
This renames VAL_BIND* routines to the more informative and
pattern-following VAL_TYPESET* names.  It also makes the bits
access of typeset more obvious with VAL_TYPESET_BITS instead
of just using VAL_TYPESET, which will also generalize better for
a typeset that may be capable of storing its list another way.

The BUF_WORDS buffer was used both for collecting keys of objects
(once unwords, now typesets) as well as in the Collect_Words routines
where they are still collecting actual REB_WORDs.  It doesn't make
sense to have two such buffers just because of the type difference
when they don't overlap, hence this is changed to BUF_COLLECT to
indicate appropiate type-neutrality.

Also, one user-visible error that should have been an assert was
converted to an assert.