Better debug alert when freeing already freed series
When not using the core routine for freeing a series, this could show
up as an assert about not being able to find a series in the manuals
tracking list.  Since freeing an already freed series can happen for
a few reasons, the common problem is better identified before giving
that potentially confusing assert.