Add very basic GC hook mechanism
This is a simple way of setting a global function that takes a void
and returns a void...but can make some number of calls to mark
REBVALs for the GC.  It also adds a typeset for determining which
values need to be tracked in this way.

While by no means the final say in Rebol extensibility mechanisms,
it's enough to keep Ren/C++ and Ren Garden from crashing after
a GC happens.