Add <stdlib.h> to the basics included before host-lib.h
There's no formal baseline for what is legal in parameter types for
host library functions... but the allocator uses size_t.  Since that is
a dependency, it means that code using host-lib.h won't compile
unless size_t is defined.

This decision could be revisited to use some other type, e.g.
REBCNT or REBU64.  But given that the idea of the host kit is
that the protocols are compiled with Rebol natively (e.g. routines
using either wchar_t or char), that decision should probably be
made when the host-and-Rebol protocols are nailed down in
other ways based on what the rules are.